+++ packages/migrate/uglifyJsPlugin/uglifyJsPlugin.test.js
@@ -5,3 +5,5 @@ const defineTest = require("webpack-cli-utils//defineTest");
 defineTest(__dirname, "uglifyJsPlugin", "uglifyJsPlugin-0");
 defineTest(__dirname, "uglifyJsPlugin", "uglifyJsPlugin-1");
 defineTest(__dirname, "uglifyJsPlugin", "uglifyJsPlugin-2");
+defineTest(__dirname, "uglifyJsPlugin", "uglifyJsPlugin-3");
+defineTest(__dirname, "uglifyJsPlugin", "uglifyJsPlugin-4");